Refactor sync methods into their own service
