+++ src/nsls2api/services/background_service.py
@@ -7,7 +7,7 @@ import bson
 
 from nsls2api.infrastructure.logging import logger
 from nsls2api.models.jobs import BackgroundJob, JobActions, JobStatus, JobSyncParameters
-from nsls2api.services import facility_service, proposal_service
+from nsls2api.services import sync_service 
 
 
 async def create_background_job(
@@ -108,35 +108,35 @@ async def worker_function():
                     logger.info(
                         f"Processing job {job.id} to update cycle information for the {job.sync_parameters.facility} facilty (from {job.sync_parameters.sync_source})."
                     )
-                    await proposal_service.worker_update_proposal_to_cycle_mapping(
+                    await sync_service.worker_update_proposal_to_cycle_mapping(
                         job.sync_parameters.facility, job.sync_parameters.cycle, job.sync_parameters.sync_source
                     )
                 case JobActions.synchronize_cycles:
                     logger.info(
                         f"Processing job {job.id} to synchronize cycles for the {job.sync_parameters.facility} facilty (from {job.sync_parameters.sync_source})."
                     )
-                    await facility_service.worker_synchronize_cycles_from_pass(
+                    await sync_service.worker_synchronize_cycles_from_pass(
                         job.sync_parameters.facility
                     )
                 case JobActions.synchronize_proposal:
                     logger.info(
                         f"Processing job {job.id} to synchronize proposal {job.sync_parameters.proposal_id} (from {job.sync_parameters.sync_source})."
                     )
-                    await proposal_service.worker_synchronize_proposal_from_pass(
+                    await sync_service.worker_synchronize_proposal_from_pass(
                         job.sync_parameters.proposal_id
                     )
                 case JobActions.synchronize_proposals_for_cycle:
                     logger.info(
                         f"Processing job {job.id} to synchronize proposals for cycle {job.sync_parameters.cycle} (from {job.sync_parameters.sync_source})."
                     )
-                    await proposal_service.worker_synchronize_proposals_for_cycle_from_pass(
+                    await sync_service.worker_synchronize_proposals_for_cycle_from_pass(
                         job.sync_parameters.cycle
                     )
                 case JobActions.synchronize_proposal_types:
                     logger.info(
                         f"Processing job {job.id} to synchronize proposal types for the {job.sync_parameters.facility} facilty (from {job.sync_parameters.sync_source})."
                     )
-                    await proposal_service.worker_synchronize_proposal_types_from_pass(
+                    await sync_service.worker_synchronize_proposal_types_from_pass(
                         job.sync_parameters.facility
                     )
                 case JobActions.create_slack_channel:
